add easylogging++

+++ coreTest/coreTest.h
@@ -2,5 +2,5 @@
 
 extern "C" __declspec(dllexport) int mul(int a, int b);
 
-float add(float x, float y);
+double add(float x, float y);
